\name{ARN82}
\alias{ARN82.eB.apr77}

\docType{data}

\title{Biomass of marine animals in colonizing experiments in Baltic Sea}

\description{
A numeric vector of biomass of species of benthic marine invertebrates that colonized experimental containers laid in the Baltic Sea.
}

\usage{
data(ARN82.eB.apr77)
}

\format{
  Named vector of positive numbers. Labels are a sequential numeric code for species, in the same sequence as the original table.
}

\details{
Each value is the biomass of invertebrate species sampled in 13
containers laid in the seafloor,as reported in Table I in Arntz & Rumohr
(1982). After the first year of exposition, 13 containers were recovered
every two months, starting in December 1976. Data are from containers recovered in April 1977.
}

\source{
Arntz, W. E., & Rumohr, H. (1982). An experimental study of macrobenthic
colonization and succession, and the importance of seasonal variation in
temperate latitudes. Journal of experimental marine biology and ecology, 64(1), 17-45.
}

\references{
Hughes, R. (1986). Theories and models of species abundance. The American Naturalist, 128(6), 879-899.
}

\keyword{datasets}
