% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.fitrad.R
\name{as.fitrad}
\alias{as.fitrad}
\title{fitsad to fitrad conversion}
\usage{
as.fitrad(object)
}
\arguments{
\item{object}{A \code{link{fitsad}} object to be converted}
}
\description{
Converts a \code{\link{fitsad}} object into a \code{\link{fitrad}} object, for comparisons
}
\details{
This function takes a fitsad object and generates the corresponding fitrad, in order to enable
comparisons such as the AIC.
}
\examples{
f1 <- fitlnorm(moths)
logLik(f1)
# This is incomparable with fitrad objects, such as
fz <- fitzipf(moths)
logLik(fz)
# However, the converted fitrad has a comparable log-likelihood
f2 <- as.fitrad(f1)
logLik(f2)
}

