\name{bci}
\docType{data}
\alias{bci}
\title{
  Tree species abundance in Barro Colorado Island Plot
}

\description{
  Number of species of trees with diameter at breast height
  >= 10 mm in a 50 ha plot of tropical forest in the Barro Colorado
  Island, Panama.
}

\usage{
data(bci)
}

\format{
  A named vector of 225 integers.
}

\details{
  The BCI forest dynamics research project was made possible by National
  Science Foundation grants to Stephen P. Hubbell: DEB-0640386,
  DEB-0425651, DEB-0346488, DEB-0129874, DEB-00753102, DEB-9909347,
  DEB-9615226, DEB-9615226, DEB-9405933, DEB-9221033, DEB-9100058,
  DEB-8906869, DEB-8605042, DEB-8206992, DEB-7922197, support from the
  Center for Tropical Forest Science, the Smithsonian Tropical Research
  Institute, the John D. and Catherine T. MacArthur Foundation, the
  Mellon Foundation, the Small World Institute Fund, and numerous
  private individuals, and through the hard work of over 100 people from
  10 countries over the past two decades. The plot project is part the
  Center for Tropical Forest Science, a global network of large-scale
  demographic tree plots. 
}

\source{
  Datasheet 'full matrix' of supplemental table 1 in Condit et al. (2002). 
}


\references{
  Condit et. al 2002. Beta-diversity in tropical forest trees. \emph{Science 295}:666-669

  Hubbell, S.P., Condit, R., and Foster, R.B. 2005. Barro Colorado
  Forest Census Plot Data. URL
  https://ctfs.arnarb.harvard.edu/webatlas/datasets/bci.

  Condit, R. 1998. \emph{Tropical Forest Census Plots}. Springer-Verlag
  and R. G. Landes Company, Berlin, Germany, and Georgetown, Texas. 
}

\keyword{datasets}
