\name{birds}
\docType{data}
\alias{birds}


\title{
  Abundances of breeding birds in Quacker Run Valley, NY
}

\description{
  Number of sights of nesting birds by species in a census
  in a 16,697 acres Park in New York State, ran in 1936 by A. A. Saunders.
}

\usage{
data(birds)
}

\format{
  A unnamed vector of integers
}

\details{
  This is one of the datasets that F. W. Preston used  to show the
  application of the lognormal distribution to describe the abundance of species
  in a sample of a biological community.
}

\source{
  Table IA of Preston (1948), which did not provide species names.
}


\references{
  Preston, F. W. 1948. The Commonness, and rarity, of
  species. \emph{Ecology 29}:254-283.

  Saunders, A. A. 1936. Studies of breeding birds in the Allegany State
  Park. New York State Museum Handbook 16. Albany, NY.
}

\keyword{datasets}
