\name{okland}
\docType{data}
\alias{okland}
\title{
  Abundances of land snail species in Norway
}

\description{
  Number of individuals of 21 species of land snails
  caught in the vegetation in Norway, 1927. 
}

\usage{
data(okland)
}

\format{
  A unnamed vector of 21 integers.
}

\details{
  This is one of the data sets that Isao Motomura used to demonstrate his
  \sQuote{law of geometric series} of the abundance of species
  in a sample of a biological community. The abundance values and fitted
  series are in figure 2 of Motomura (1932).
}

\source{
  Second column ('A') of table 5 of Okland (1930).
}


\references{

  Motomura, I. 1932. On the statistical treatment of communities
  (in Japanese).  \emph{Zool. Mag. (Tokyo) 44}: 379--383.  English translation in Appendix
  I in Doi, H. and Mori, T. 2012. The discovery of species-abundance
  distribution in an ecological community. \emph{Oikos 122}: 179--182.
  
  Okland, F. 1930. Quantitative Untersuchungen der Landschneckenfauna
  Norwegens. I. \emph{Zoomorphology 16:} 748--804.

}

\keyword{datasets}
