% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sads-methods.R
\name{coefficients,fitsad-method}
\alias{coefficients,fitsad-method}
\alias{coefficients,fitsadC-method}
\alias{coefficients,fitrad-method}
\alias{fitted.values,fitsad-method}
\alias{fitted.values,fitsadC-method}
\alias{fitted.values,fitrad-method}
\alias{fitted,fitsad-method}
\alias{fitted,fitsadC-method}
\alias{fitted,fitrad-method}
\alias{residuals,fitsad-method}
\alias{residuals,fitsadC-method}
\alias{residuals,fitrad-method}
\title{Standard stats methods}
\usage{
\S4method{coefficients}{fitsad}(object, ...)

\S4method{coefficients}{fitsadC}(object, ...)

\S4method{coefficients}{fitrad}(object, ...)

\S4method{fitted.values}{fitsad}(object, ...)

\S4method{fitted.values}{fitsadC}(object, ...)

\S4method{fitted.values}{fitrad}(object, ...)

\S4method{fitted}{fitsad}(object, ...)

\S4method{fitted}{fitsadC}(object, ...)

\S4method{fitted}{fitrad}(object, ...)

\S4method{residuals}{fitsad}(object, ...)

\S4method{residuals}{fitsadC}(object, ...)

\S4method{residuals}{fitrad}(object, ...)
}
\arguments{
\item{object}{An object from class fitsad, fitrad or fitsadC}

\item{\dots}{Other arguments to be forwarded for the lower level function}
}
\description{
Provide the standard interface for fitted objects
}
\details{
These methods are provided to allow for standard manipulation of \code{\link{fitsad}}, \code{\link{fitsadC}} 
and \code{\link{fitrad}} objects using the generic methods defined in the "stats" package.
Please see the original man pages for each method.

\code{coefficients} is an alias to \code{\link[stats]{coef}} (implemented in package "bbmle").

\code{fitted} and \code{fitted.values} provide an alternative interface to \code{\link{radpred}};
these are also used to calcutate \code{residuals}.

Notice that radpred is a preferred interface for most calculations, specially if there are several
ties.
}
