% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sads-classes.R, R/sads-methods.R
\docType{class}
\name{summary.sads-class}
\alias{summary.sads-class}
\alias{show,summary.sads-method}
\alias{summary,fitsad-method}
\alias{summary,fitrad-method}
\title{Summary for fitsad/fitrad calls}
\usage{
\S4method{show}{summary.sads}(object)

\S4method{summary}{fitsad}(object)

\S4method{summary}{fitrad}(object)
}
\arguments{
\item{object}{An object of class fitsad/fitrad is required to generate a summary.sads object.}
}
\description{
This function works almost exactly as bbmle's summary.mle2, but it includes a "fixed parameters" 
line for models with fixed parameters, such as \code{\link{fitls}} or \code{fitvolkov}.
}
