% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distr.R
\name{distr}
\alias{distr}
\title{Continuous or discrete distributions}
\usage{
distr(distribution)
}
\arguments{
\item{distribution}{Character. The name of the distribution ("geom" for "fitgeom", "weibull" for "fitweibull", etc.}
}
\description{
Checks if the distribution is continuous or discrete
}
\details{
Returns whether a given distribution (used for a sad or rad model) is "discrete"
or "continuous". The name is compared to a list of known distributions, so distributions
not used in the sads package will return "NA" with a warning.

In the package sads up to version 0.2.3, the user was required to explicitly set
a distr argument in some calls to \code{\link{radpred}} and \code{\link{qqsad}}. Now
this is handled automatically, and attempts to set the "distr" argument explicitly are ignored.
}

