\name{rsad}
\alias{rsad}

\title{Poisson and negative binomial sampling of a species abundance distribution
}

\description{
  A given number of realizations of a probability distribution (species
  abundances in a community) is sampled by a Poisson or Negative Binomial process. 
}

\usage{
rsad(S = NULL, frac, sad, coef, trunc = NaN, 
     sampling = c("poisson", "nbinom", "fixed"), k , zeroes = FALSE, ssize=1)
}

\arguments{
  \item{S}{
    positive integer; number of species in the community, which is the number of random deviates
    generated by the probability distribution given by argument \code{sad}. Notice that for some
    distributions, the value of \code{S} is deduced from the coefficients, so this value is ignored
    (see details).
  }
  \item{frac}{
    single numeric \code{0 < frac <= 1}; fraction of the community sampled. Notice that the
    assumptions behind this function are only valid for small values of \code{frac}.
  }
  \item{sad}{
    character; root name of community sad distribution - e.g., \kbd{lnorm}
    for the lognormal distribution \code{rlnorm}; \kbd{geom} for the geometric distribution
	\code{rgeom} or \kbd{ls} for the lognormal distribution \code{\link{ls}}.
  }  
  \item{coef}{
    list with named arguments to be passed to the probability function defined by the
    argument sad.
  }
  \item{trunc}{
    The truncation point at which the random distribution should be truncated; see \code{\link{rtrunc}}.
    The default value of \code{NaN} means that no truncation is performed.
  }
  \item{sampling}{
    character; if \kbd{poisson} the sampling process is Poisson (independent sampling of
    individuals); if \kbd{nbinom} negative binomial
    sampling is used to simulate aggregation of individuals in sampling units;
    finally, \kbd{fixed} samples a fixed number of \code{frac} times the number of 
    individuals in the simulated community, without replacement. Partial matching is allowed.
  }
  \item{k}{
    positive; size parameter for the sampling binomial negative. This parameter is ignored for other sampling
    techniques.
  }
  \item{zeroes}{
    logical; should zero values be included in the returned vector?
  }
  \item{ssize}{
    positive integer; sample size: number of draws taken from the community.
  }
}

\details{
  This function simulates one or more random samples taken from a community with \code{S}
  species. The expected species abundances in the sampled community follow a
  probability distribution given by the argument \code{sad}. A fraction \code{frac} of
  the community is sampled, thus the expected abundance in the sample of each
  species is \code{frac*n}, where n is the species' expected abundance in the
  community.

  Three sampling processes can be simulated: Poisson,
  where individuals are sampled independently, and negative binomial,
  where individuals are aggregated over sampling units. Finally, the "fixed"
  sampling scheme takes a sample of \code{frac * n} individuals, without replacement.

  In general terms, this function takes a Poisson or negative
  binomial sampling with replacement of a vector of \code{S} realizations of a random variable,
  with the sampling intensity given by \code{frac}. The resulting values are
  realizations of a Poisson (or a Negative Binomial) random variable where the
  parameter that corresponds to the mean (=expected value of the variable) follows a probability
  distribution given by the argument \code{sad}.

  For the broken-stick, logseries, MZSM and Volkov distributions, the expected value of \code{S}
  is deduced from the coefficients provided in the argument \code{coef}; thus, the value of the parameter
  \code{S} is ignored and may be left blank. The formulas for the number of species in each case are:

  * Broken-stick: coefficient S
  * Log-series: alpha log(1 + N/alpha)
  * MZSM: sum_x=1^J theta/x (1 - x/J)^(theta - 1)
  * Volkov: sum of the unnormalized PDF from 1 to J, see \code{\link{dvolkov}}
}

\value{
  if \kbd{ssize=1} a vector of (zero truncated) abundances in the sample;
  if \kbd{ssize>1} a data frame with sample identification, species identification, and (zero truncated) abundances.
}

\references{
  Pielou, E.C. 1977. \emph{Mathematical Ecology}. New York: John Wiley
  and Sons.
  
  Green,J. and Plotkin, J.B. 2007 A statistical theory for sampling species abundances. \emph{Ecology Letters 10}:1037--1045
}

\author{
  Paulo I. Prado \email{prado@ib.usp.br} and Andre Chalom.
}


\seealso{
  \code{\link{dpoix}}, \code{\link{dpoig}} and \code{\link{dpoilog}} for
  examples of compound Poisson probability distributions like those
  simulated by \code{rsad}.
}

\examples{

##A Poisson sample from a community with a lognormal sad
samp2 <- rsad(S = 100, frac=0.1, sad="lnorm", coef=list(meanlog=5, sdlog=2))
## Preston plot
plot(octav(samp2))
## Once this is a Poisson sample of a lognormal community, the abundances
## in the sample should follow a Poisson-lognormal distribution.
## Adds line of theoretical Poisson-lognormal with
## mu=meanlog+log(frac) and sigma=sdlog)
## Predicted by the theoretical Poisson-lognormal truncated at zero
samp2.pred <- octavpred(samp2, sad="poilog", coef= list(mu=5+log(0.1), sig=2), trunc=0)
## Adding the line in the Preston plot
lines(samp2.pred)
}
