\name{rsad}
\alias{rsad}
\title{Poisson and Negative Binomial Sampling of a community sad
}
\description{
A given number of realizations of a probability distribution (species
abundances in a community) is sampled by a Poisson or Negative Binomial process. 
}
\usage{
rsad(S, frac, sad, samp=c("Poisson","Negbinom"), k , zeroes=FALSE,
ssize=1, ...)
}
\arguments{
  \item{S}{
    positive integer; number of species in the community, which is the number of random deviates
    generated by the probability distribution given by argument 'sad'
  }
  \item{frac}{
    single numeric '0<frac<=1'; fraction of the community sampled
  }
  \item{sad}{
    character; root name of community sad distribution - e.g., "lnorm"
    for the lognormal distribution; 
    "geom" for the geometric distribution.
  }
  
  \item{samp}{
    character; sampling process: "Poisson" (independent sampling of
    individuals)
    or "Negbinom" for negative binomial
    sampling to simulate aggregation of individuals in sampling units.
  }
  \item{k}{
    positive; size parameter for the sampling binomial negative 
  }
  \item{zeroes}{
    logical; should zero values be included in the returned vector?
  }
  \item{ssize}{
    integer>0; sample size: number of draws taken from the community.
    }
  \item{...}{
    named arguments to be passed to the probability function defined by the
    argument sad.
  }
}
\details{
This function simulates a random sample taken from a community with S
species. The expected species abundances in the sampled community follow a
probability distribution given by the argument 'sad'. A fraction 'frac' of
the community is sampled, thus the expected abundance in the sample of each
species is frac*n, where n is the species' expected abundance in the
community.

Two sampling processes can be simulated: a Poisson sample,
where individuals are sampled independently, and negative binomial
sample, where individuals are aggregated over sampling units.

In general terms, this function takes a Poisson or negative
binomial sampling with reposition of a vector of S realizations of a random variable,
with the sampling intensity given by 'frac'. The resulting values are
realizations of a Poisson (or a Negative Binomial) random variable where the
parameter that correspond to the mean (=expected value of the variable) follows a probability
distribution given by the argument 'sad'.
}

\value{
  if 'ssize'=1 a vector of (zero truncated) abundances in the sample;
  if 'ssize'>1 a data.frame with sample identification, species identification, and (zero truncated) abundances.
}

\references{

  Pielou, E.C. 1977. \emph{Mathematical Ecology}. New York: John Wiley
  and Sons.
  
  Green,J. and Plotkin, J.B. 2007 A statistical theory for sampling species abundances. \emph{Ecology Letters 10}:1037--1045
}

\author{
Paulo I. Prado
}


\seealso{
dexp, dpois, dpoix
}

\examples{

##A Poisson sample from a community with an exponential sad
exp.samp <- rsad(S=200,frac=0.05,"exp",rate=1/250)
hist(exp.samp, prob=TRUE, xlab="Abundance in the sample")
x<-0:max(exp.samp)
y<-dtrunc("poix",x,trunc=0,coef=list(frac=0.05, rate=1/250)))## theoretical
distribution truncated at zero
lines(x,y, col="blue", type="b", lty=2)
}
