\name{sads-package}
\alias{sads-package}
\alias{sads}
\docType{package}
\title{
Modeling Species Abundance Distributions 
}
\description{
Likelihood-based fit and selection of models for species abundance distributions
}
\details{
\tabular{ll}{
Package: \tab sads\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-02-02\cr
License: \tab gpl \cr
LazyLoad: \tab yes\cr
}

Tools to fit hierachical models for species abundance
distributions (sads). These models include the theoretical sads in the
community and the sampling process that generated the observed abundance values.

}
\author{
Paulo I Prado and Murilo Dantas Miranda

Maintainer: P.I. Prado  <prado@ib.usp.br>
}

\references{
Green,J. and Plotkin, J.B. 2007 A statistical theory for sampling
species abundances. \emph{Ecology Letters 10}:1037--1045
}

\keyword{ species abundance distributions, ecology, environmetrics,
  maximum likelihood estimation }

\seealso{
vegan-package, poilog-package, untb-package
}

