\name{sads-package}
\alias{sads-package}
\alias{sads}
\docType{package}
\title{
What the package does (short line)
species
}
\description{
More about what it does (maybe more than one line)
x
}
\details{
\tabular{ll}{
Package: \tab sads\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-24\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
x
}
\references{
x
}
